Add meta field in routing paths and refactor navigation.vue
Added a meta field in routing paths of main.js for 'month' and 'week' components. In navigation.vue, the code was refactored to remove computed properties dependencies on state. Instead, the current date is now directly calculated and a function isActive, aligning with the added meta, was implemented to indicate the current active route. Also, removed unnecessary comment in month.vue.